+++ examples/cli_client.rs
@@ -48,7 +48,7 @@ async fn main() -> std::result::Result<(), anyhow::Error> {
 
 async fn process_input(conn: &Connection, line: String) {
     let query = format!("return ({})", line);
-    match conn.eval::<_, serde_json::Value>(query, vec![]).await {
+    match conn.eval::<serde_json::Value, _, _>(query, ()).await {
         Ok(x) => println!(
             "Result: {}",
             serde_json::to_string(&x).expect("All MessagePack values should be valid for JSON")