+++ src/client/executor_ext.rs
@@ -11,6 +11,7 @@ use crate::{
         Call, Delete, EncodedRequest, Eval, Insert, Ping, Replace, Request, Select, Update, Upsert,
     },
     schema::{SchemaEntityKey, Space},
+    tuple::Tuple,
     utils::extract_and_deserialize_iproto_data,
     IteratorType, Result,
 };
@@ -37,10 +38,11 @@ pub trait ExecutorExt: Executor {
     /// Evaluate Lua expression.
     ///
     /// Check [docs][crate#deserializing-lua-responses-in-call-and-eval] on how to deserialize response.
-    async fn eval<I, T>(&self, expr: I, args: Vec<Value>) -> Result<T>
+    async fn eval<T, A, I>(&self, expr: I, args: A) -> Result<T>
     where
-        I: Into<Cow<'static, str>> + Send,
         T: DeserializeOwned,
+        A: Tuple + Send,
+        I: Into<Cow<'static, str>> + Send,
     {
         let body = self.send_request(Eval::new(expr, args)).await?;
         extract_and_deserialize_iproto_data(body).map_err(Into::into)
@@ -49,27 +51,29 @@ pub trait ExecutorExt: Executor {
     /// Remotely call function in Tarantool.
     ///
     /// Check [docs][crate#deserializing-lua-responses-in-call-and-eval] on how to deserialize response.
-    async fn call<I, T>(&self, function_name: I, args: Vec<Value>) -> Result<T>
+    async fn call<T, A, I>(&self, function_name: I, args: A) -> Result<T>
     where
-        I: Into<Cow<'static, str>> + Send,
         T: DeserializeOwned,
+        A: Tuple + Send,
+        I: Into<Cow<'static, str>> + Send,
     {
         let body = self.send_request(Call::new(function_name, args)).await?;
         extract_and_deserialize_iproto_data(body).map_err(Into::into)
     }
 
     /// Select tuples from space.
-    async fn select<T>(
+    async fn select<T, A>(
         &self,
         space_id: u32,
         index_id: u32,
         limit: Option<u32>,
         offset: Option<u32>,
         iterator: Option<IteratorType>,
-        keys: Vec<Value>,
+        keys: A,
     ) -> Result<Vec<T>>
     where
         T: DeserializeOwned,
+        A: Tuple + Send,
     {
         let body = self
             .send_request(Select::new(
@@ -81,32 +85,34 @@ pub trait ExecutorExt: Executor {
 
     // TODO: decode response
     /// Insert tuple.
-    async fn insert(&self, space_id: u32, tuple: Vec<Value>) -> Result<()> {
+    async fn insert<T>(&self, space_id: u32, tuple: T) -> Result<()>
+    where
+        T: Tuple + Send,
+    {
         let _ = self.send_request(Insert::new(space_id, tuple)).await?;
         Ok(())
     }
 
-    // TODO: structured tuple
     // TODO: decode response
     /// Update tuple.
-    async fn update(
-        &self,
-        space_id: u32,
-        index_id: u32,
-        keys: Vec<Value>,
-        tuple: Vec<Value>,
-    ) -> Result<()> {
+    async fn update<K, T>(&self, space_id: u32, index_id: u32, keys: K, tuple: T) -> Result<()>
+    where
+        K: Tuple + Send,
+        T: Tuple + Send,
+    {
         let _ = self
             .send_request(Update::new(space_id, index_id, keys, tuple))
             .await?;
         Ok(())
     }
 
-    // TODO: structured tuple
     // TODO: decode response
-    // TODO: maybe set index base to 1 always?
     /// Update or insert tuple.
-    async fn upsert(&self, space_id: u32, ops: Vec<Value>, tuple: Vec<Value>) -> Result<()> {
+    async fn upsert<O, T>(&self, space_id: u32, ops: O, tuple: T) -> Result<()>
+    where
+        O: Tuple + Send,
+        T: Tuple + Send,
+    {
         let _ = self.send_request(Upsert::new(space_id, ops, tuple)).await?;
         Ok(())
     }
@@ -115,7 +121,10 @@ pub trait ExecutorExt: Executor {
     // TODO: decode response
     /// Insert a tuple into a space. If a tuple with the same primary key already exists,
     /// replaces the existing tuple with a new one.
-    async fn replace(&self, space_id: u32, keys: Vec<Value>) -> Result<()> {
+    async fn replace<T>(&self, space_id: u32, keys: T) -> Result<()>
+    where
+        T: Tuple + Send,
+    {
         let _ = self.send_request(Replace::new(space_id, keys)).await?;
         Ok(())
     }
@@ -123,7 +132,10 @@ pub trait ExecutorExt: Executor {
     // TODO: structured tuple
     // TODO: decode response
     /// Delete a tuple identified by the primary key.
-    async fn delete(&self, space_id: u32, index_id: u32, keys: Vec<Value>) -> Result<()> {
+    async fn delete<T>(&self, space_id: u32, index_id: u32, keys: T) -> Result<()>
+    where
+        T: Tuple + Send,
+    {
         let _ = self
             .send_request(Delete::new(space_id, index_id, keys))
             .await?;
@@ -203,7 +215,7 @@ mod ui {
     fn executor_ext_on_transaction() {
         async fn f(tx: Transaction) {
             let space_tx: Space<Transaction> = tx.into_space("space").await.unwrap().unwrap();
-            space_tx.delete(vec![1.into()]).await.unwrap();
+            space_tx.delete((1,)).await.unwrap();
             space_tx.commit().await.unwrap();
         }
     }