+++ src/codec/request/call.rs
@@ -1,26 +1,23 @@
-// TODO: unify with eval.rs
-
 use std::{borrow::Cow, io::Write};
-
-use rmpv::Value;
 
 use crate::{
     codec::{
         consts::{keys, RequestType},
-        utils::{write_kv_array, write_kv_str},
+        utils::{write_kv_str, write_kv_tuple},
     },
     errors::EncodingError,
+    tuple::Tuple,
 };
 
 use super::Request;
 
 #[derive(Clone, Debug)]
-pub(crate) struct Call {
+pub(crate) struct Call<T> {
     pub function_name: Cow<'static, str>,
-    pub tuple: Vec<Value>,
+    pub tuple: T,
 }
 
-impl Request for Call {
+impl<T: Tuple> Request for Call<T> {
     fn request_type() -> RequestType
     where
         Self: Sized,
@@ -32,14 +29,13 @@ impl Request for Call {
     fn encode(&self, mut buf: &mut dyn Write) -> Result<(), EncodingError> {
         rmp::encode::write_map_len(&mut buf, 2)?;
         write_kv_str(buf, keys::FUNCTION_NAME, self.function_name.as_ref())?;
-        write_kv_array(buf, keys::TUPLE, &self.tuple)?;
+        write_kv_tuple(buf, keys::TUPLE, &self.tuple)?;
         Ok(())
     }
 }
 
-impl Call {
-    // TODO: introduce some convenient way to pass arguments
-    pub(crate) fn new(function_name: impl Into<Cow<'static, str>>, args: Vec<Value>) -> Self {
+impl<T> Call<T> {
+    pub(crate) fn new(function_name: impl Into<Cow<'static, str>>, args: T) -> Self {
         Self {
             function_name: function_name.into(),
             tuple: args,