+++ src/codec/request/delete.rs
@@ -1,26 +1,25 @@
 use std::io::Write;
 
-use rmpv::Value;
-
 use crate::{
     codec::{
         consts::{keys, RequestType},
-        utils::{write_kv_array, write_kv_u32},
+        utils::{write_kv_tuple, write_kv_u32},
     },
     errors::EncodingError,
+    tuple::Tuple,
 };
 
 use super::Request;
 
 #[derive(Clone, Debug)]
-pub(crate) struct Delete {
+pub(crate) struct Delete<T> {
     pub space_id: u32,
     pub index_id: u32,
-    pub keys: Vec<Value>,
+    pub keys: T,
 }
 
-impl Delete {
-    pub(crate) fn new(space_id: u32, index_id: u32, keys: Vec<Value>) -> Self {
+impl<T> Delete<T> {
+    pub(crate) fn new(space_id: u32, index_id: u32, keys: T) -> Self {
         Self {
             space_id,
             index_id,
@@ -29,7 +28,7 @@ impl Delete {
     }
 }
 
-impl Request for Delete {
+impl<T: Tuple> Request for Delete<T> {
     fn request_type() -> RequestType
     where
         Self: Sized,
@@ -42,7 +41,7 @@ impl Request for Delete {
         rmp::encode::write_map_len(&mut buf, 3)?;
         write_kv_u32(buf, keys::SPACE_ID, self.space_id)?;
         write_kv_u32(buf, keys::INDEX_ID, self.index_id)?;
-        write_kv_array(buf, keys::KEY, &self.keys)?;
+        write_kv_tuple(buf, keys::KEY, &self.keys)?;
         Ok(())
     }
 }