+++ src/codec/utils.rs
@@ -1,8 +1,6 @@
 use std::io::Write;
 
-use rmpv::Value;
-
-use crate::errors::EncodingError;
+use crate::{errors::EncodingError, tuple::Tuple};
 
 pub fn write_kv_str(mut buf: &mut dyn Write, key: u8, value: &str) -> Result<(), EncodingError> {
     rmp::encode::write_pfix(&mut buf, key)?;
@@ -16,16 +14,12 @@ pub fn write_kv_u32(mut buf: &mut dyn Write, key: u8, value: u32) -> Result<(), 
     Ok(())
 }
 
-pub fn write_kv_array(
+pub fn write_kv_tuple<T: Tuple>(
     mut buf: &mut dyn Write,
     key: u8,
-    value: &[Value],
+    tuple: T,
 ) -> Result<(), EncodingError> {
     rmp::encode::write_pfix(&mut buf, key)?;
-    // TODO: safe conversion from usize to u32
-    rmp::encode::write_array_len(&mut buf, value.len() as u32)?;
-    for x in value.iter() {
-        rmpv::encode::write_value(&mut buf, x)?;
-    }
+    T::encode_into_writer(&tuple, buf)?;
     Ok(())
 }