+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
  - `into_space` method to `ExecutorExt` trait, wich return `Space` with underlying `Executor`;
  - `.commit()` and `.rollback()` methods to `Space<Transaction>` and `OwnedIndex<Transaction>`;
- - `timeout` parameter to `ConnectionBuilder`, allowing to set timeout for all requests in this `Connection`.
+ - `timeout` parameter to `ConnectionBuilder`, allowing to set timeout for all requests in this `Connection`;
+ - `Tuple` trait for passing arguments to requests.
 
 ### Changed
  - `get_space` moved to `ExecutorExt` trait and renamed to `space`, also now returning reference to underlying `Execitor`.